backend/requirements.txt
APScheduler==3.10.1
beautifulsoup4==4.12.2
blinker==1.6.2
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
Flask==2.3.1
Flask-Cors==3.0.10
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
psycopg2-binary==2.9.6
python-dotenv==1.0.0
python-http-client==3.3.7
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
requests==2.29.0
sendgrid==6.10.0
six==1.16.0
soupsieve==2.4.1
SQLAlchemy==2.0.11
starkbank-ecdsa==2.2.0
typing_extensions==4.5.0
tzdata==2023.3
tzlocal==4.3
urllib3==1.26.15
Werkzeug==2.3.2
